requirements.txt
aiofiles==24.1.0
aiogram==3.18.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.13
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
async-timeout==5.0.1
attrs==25.2.0
CacheControl==0.14.3
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.2
click==8.1.8
cryptography==44.0.3
dotenv==0.9.9
exceptiongroup==1.2.2
fastapi==0.115.12
filelock==3.18.0
firebase-admin==6.8.0
frozenlist==1.5.0
fsspec==2025.3.2
google-api-core==2.24.2
google-api-python-client==2.169.0
google-auth==2.39.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.3
google-cloud-firestore==2.20.2
google-cloud-storage==3.1.0
google-crc32c==1.7.1
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
grpcio==1.71.0
grpcio-status==1.71.0
h11==0.16.0
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
idna==3.10
Jinja2==3.1.6
llvmlite==0.44.0
magic-filter==1.0.12
MarkupSafe==3.0.2
more-itertools==10.7.0
mpmath==1.3.0
msgpack==1.1.0
multidict==6.1.0
networkx==3.4.2
numba==0.61.2
numpy==1.26.4
openai-whisper==20240930
pillow==11.2.1
propcache==0.3.0
proto-plus==1.26.1
protobuf==5.29.4
pyasn1==0.6.1
pyasn1_modules==0.4.2
pybind11==2.13.6
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
pydub==0.25.1
PyJWT==2.10.1
pyparsing==3.2.3
PyPDF2==3.0.1
python-dotenv==1.1.0
python-multipart==0.0.20
python-telegram-bot==22.0
regex==2024.11.6
reportlab==4.4.0
requests==2.32.3
rsa==4.9.1
shortuuid==1.0.13
sniffio==1.3.1
starlette==0.46.2
sympy==1.14.0
tiktoken==0.9.0
torch==2.2.2
tqdm==4.67.1
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.4.0
uvicorn==0.34.2
yarl==1.18.3
